+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 =========
 
 
+v0.4.0 (2015-05-12)
+-------------------
+
 - Make ``Between`` accept keyword arguments for ``min`` and ``max``.
 - Make ``Length`` function like ``Between`` and allow comparison over range of lengths. If a single comparable value is passed in, then comparison uses the value as a max length. Previously, a single comparable value performed an equality check for length. (**breaking change**)
 - Make ``Match`` accept keyword argument ``flags`` for use with string based regular expression.